Coding: Spotlight

+++ OpenGLCourseApp/OpenGLCourseApp/PointLight.cpp
@@ -26,9 +26,9 @@ PointLight::~PointLight()
 {
 }
 
-void PointLight::UseLight(GLfloat ambientIntensityLocation, GLuint ambientColourLocation, GLuint diffuseIntensityLocation, 
-	GLuint positionLocation, 
-	GLfloat constantLocation, GLfloat linearLocation, GLfloat exponentLocation)
+void PointLight::UseLight(GLuint ambientIntensityLocation, GLuint ambientColourLocation,
+	GLuint diffuseIntensityLocation, GLuint positionLocation,
+	GLuint constantLocation, GLuint linearLocation, GLuint exponentLocation)
 {
 	glUniform3f(ambientColourLocation, colour.x, colour.y, colour.z);
 	glUniform1f(ambientIntensityLocation, ambientIntensity);
